Move DataTracer to core module
