+++ tests/test_core.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import numpy as np
 import pandas as pd
 
-from datatracer import DataTracer
+from datatracer.core import DataTracer
 
 
 class TestDataTracer(TestCase):
