+++ tests/test_datatracer.py
@@ -1,0 +1,22 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""Tests for `data_tracer` package."""
+
+import unittest
+
+# from datatracer import data_tracer
+
+
+class TestData_tracer(unittest.TestCase):
+    """Tests for `data_tracer` package."""
+
+    def setUp(self):
+        """Set up test fixtures, if any."""
+
+    def tearDown(self):
+        """Tear down test fixtures, if any."""
+
+    def test_000_something(self):
+        """Test something."""
+        self.assertTrue(True)
